feat(eula): rename eula.rs to licenses.rs
